Fix bug show Winner list when Buy Type = `FCFS`

+++ backend/database/migrations/1617879679199_add_field_min_buy_tiers_table_schema.js
@@ -7,7 +7,7 @@ class AddFieldMinBuyToTiersTableSchema extends Schema {
   up () {
     this.table('tiers', (table) => {
       // alter table
-      table.integer('min_buy').nullable().defaultTo(0);
+      table.float('min_buy').nullable().defaultTo(0);
     })
   }
